+++ runtime/crab/src/pallets/authority_discovery.rs
@@ -1,0 +1,6 @@
+// --- substrate ---
+use pallet_authority_discovery::Config;
+// --- darwinia ---
+use crate::*;
+
+impl Config for Runtime {}
