+++ runtime/crab/src/pallets/ethereum_backing.rs
@@ -1,0 +1,31 @@
+// --- substrate ---
+use sp_runtime::ModuleId;
+// --- darwinia ---
+use crate::*;
+use darwinia_ethereum_backing::Config;
+
+frame_support::parameter_types! {
+	pub const EthereumBackingModuleId: ModuleId = ModuleId(*b"da/ethbk");
+	pub const EthereumBackingFeeModuleId: ModuleId = ModuleId(*b"da/ethfe");
+	pub const RingLockLimit: Balance = 10_000_000 * COIN;
+	pub const KtonLockLimit: Balance = 1_000 * COIN;
+	// https://github.com/darwinia-network/darwinia-common/pull/377#issuecomment-730369387
+	pub const AdvancedFee: Balance = 50 * COIN;
+	pub const SyncReward: Balance = 1_000 * COIN;
+}
+impl Config for Runtime {
+	type ModuleId = EthereumBackingModuleId;
+	type FeeModuleId = EthereumBackingFeeModuleId;
+	type Event = Event;
+	type RedeemAccountId = AccountId;
+	type EthereumRelay = EthereumRelay;
+	type OnDepositRedeem = Staking;
+	type RingCurrency = Ring;
+	type KtonCurrency = Kton;
+	type RingLockLimit = RingLockLimit;
+	type KtonLockLimit = KtonLockLimit;
+	type AdvancedFee = AdvancedFee;
+	type SyncReward = SyncReward;
+	type EcdsaAuthorities = EthereumRelayAuthorities;
+	type WeightInfo = ();
+}
