+++ runtime/crab/src/pallets/grandpa.rs
@@ -1,0 +1,22 @@
+pub use pallet_grandpa::AuthorityId as GrandpaId;
+
+// --- substrate ---
+use frame_support::traits::KeyOwnerProofSystem;
+use pallet_grandpa::{Config, EquivocationHandler};
+use sp_core::crypto::KeyTypeId;
+// --- darwinia ---
+use crate::*;
+
+impl Config for Runtime {
+	type Event = Event;
+	type Call = Call;
+	type KeyOwnerProof =
+		<Self::KeyOwnerProofSystem as KeyOwnerProofSystem<(KeyTypeId, GrandpaId)>>::Proof;
+	type KeyOwnerIdentification = <Self::KeyOwnerProofSystem as KeyOwnerProofSystem<(
+		KeyTypeId,
+		GrandpaId,
+	)>>::IdentificationTuple;
+	type KeyOwnerProofSystem = Historical;
+	type HandleEquivocation = EquivocationHandler<Self::KeyOwnerIdentification, Offences>;
+	type WeightInfo = ();
+}
