+++ runtime/crab/src/pallets/header_mmr.rs
@@ -1,0 +1,5 @@
+// --- darwinia ---
+use crate::*;
+use darwinia_header_mmr::Config;
+
+impl Config for Runtime {}
