+++ runtime/crab/src/pallets/membership.rs
@@ -1,0 +1,28 @@
+// --- substrate ---
+use frame_support::traits::ChangeMembers;
+use pallet_membership::{Config, Instance0};
+// --- darwinia ---
+use crate::*;
+
+pub struct MembershipChangedGroup;
+impl ChangeMembers<AccountId> for MembershipChangedGroup {
+	fn change_members_sorted(
+		incoming: &[AccountId],
+		outgoing: &[AccountId],
+		sorted_new: &[AccountId],
+	) {
+		TechnicalCommittee::change_members_sorted(incoming, outgoing, sorted_new);
+		EthereumRelay::change_members_sorted(incoming, outgoing, sorted_new);
+	}
+}
+
+impl Config<Instance0> for Runtime {
+	type Event = Event;
+	type AddOrigin = EnsureRootOrMoreThanHalfCouncil;
+	type RemoveOrigin = EnsureRootOrMoreThanHalfCouncil;
+	type SwapOrigin = EnsureRootOrMoreThanHalfCouncil;
+	type ResetOrigin = EnsureRootOrMoreThanHalfCouncil;
+	type PrimeOrigin = EnsureRootOrMoreThanHalfCouncil;
+	type MembershipInitialized = TechnicalCommittee;
+	type MembershipChanged = MembershipChangedGroup;
+}
