+++ runtime/crab/src/pallets/staking.rs
@@ -1,0 +1,60 @@
+#[cfg(feature = "std")]
+pub use darwinia_staking::{Forcing, StakerStatus};
+
+// --- substrate ---
+use sp_runtime::{transaction_validity::TransactionPriority, ModuleId, Perbill};
+use sp_staking::SessionIndex;
+// --- darwinia ---
+use crate::{weights::darwinia_staking::WeightInfo, *};
+use darwinia_staking::{Config, EraIndex};
+
+frame_support::parameter_types! {
+	pub const StakingModuleId: ModuleId = ModuleId(*b"da/staki");
+	pub const SessionsPerEra: SessionIndex = SESSIONS_PER_ERA;
+	pub const BondingDurationInEra: EraIndex = 14 * DAYS
+		/ (SESSIONS_PER_ERA as BlockNumber * BLOCKS_PER_SESSION);
+	pub const BondingDurationInBlockNumber: BlockNumber = 14 * DAYS;
+	// slightly less than 14 days.
+	pub const SlashDeferDuration: EraIndex = 14 * DAYS
+		/ (SESSIONS_PER_ERA as BlockNumber * BLOCKS_PER_SESSION) - 1;
+	// quarter of the last session will be for election.
+	pub const ElectionLookahead: BlockNumber = BLOCKS_PER_SESSION / 4;
+	pub const MaxIterations: u32 = 5;
+	pub MinSolutionScoreBump: Perbill = Perbill::from_rational_approximation(5u32, 10_000);
+	pub const MaxNominatorRewardedPerValidator: u32 = 64;
+	pub const StakingUnsignedPriority: TransactionPriority = TransactionPriority::max_value() / 2;
+	pub const Cap: Balance = CAP;
+	pub const TotalPower: Power = TOTAL_POWER;
+}
+impl Config for Runtime {
+	type Event = Event;
+	type ModuleId = StakingModuleId;
+	type UnixTime = Timestamp;
+	type SessionsPerEra = SessionsPerEra;
+	type BondingDurationInEra = BondingDurationInEra;
+	type BondingDurationInBlockNumber = BondingDurationInBlockNumber;
+	type SlashDeferDuration = SlashDeferDuration;
+	/// A super-majority of the council can cancel the slash.
+	type SlashCancelOrigin = EnsureRootOrHalfCouncil;
+	type SessionInterface = Self;
+	type NextNewSession = Session;
+	type ElectionLookahead = ElectionLookahead;
+	type Call = Call;
+	type MaxIterations = MaxIterations;
+	type MinSolutionScoreBump = MinSolutionScoreBump;
+	type MaxNominatorRewardedPerValidator = MaxNominatorRewardedPerValidator;
+	type UnsignedPriority = StakingUnsignedPriority;
+	// The unsigned solution weight targeted by the OCW. We set it to the maximum possible value of
+	// a single extrinsic.
+	type OffchainSolutionWeightLimit = OffchainSolutionWeightLimit;
+	type RingCurrency = Ring;
+	type RingRewardRemainder = Treasury;
+	type RingSlash = Treasury;
+	type RingReward = ();
+	type KtonCurrency = Kton;
+	type KtonSlash = Treasury;
+	type KtonReward = ();
+	type Cap = Cap;
+	type TotalPower = TotalPower;
+	type WeightInfo = WeightInfo<Runtime>;
+}
