+++ runtime/crab/src/pallets/sudo.rs
@@ -1,0 +1,9 @@
+// --- substrate ---
+use pallet_sudo::Config;
+// --- darwinia ---
+use crate::*;
+
+impl Config for Runtime {
+	type Event = Event;
+	type Call = Call;
+}
